python/pyproject.toml
[tool.poetry]
name = "Resume_Agent_App"
version = "0.1.0"
description = "My LangChain project"
authors = ["worachot.C65"] 
readme = "README.md"
packages = [{include = "agent"}, {include = "tasks"}, {include = "auth"}, {include = "controller"}, {include = "models"}]

[tool.poetry.dependencies]
python = "3.12.10"

langchain-text-splitters = "0.3.11"
langchain-core = "0.3.79"
langchain-voyageai = "0.2.0"
langchain-mongodb = "0.7.2"
langchain-community = "0.3.17"  
langchain-anthropic = "0.3.22"
python-dotenv = "^1.0.0"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
